quickcheck-ai-meeeee/backend/requirements.txt
fastapi
uvicorn[standard]
faster-whisper
webrtcvad-wheels
numpy
